+++ src/main/java/com/heartsave/todaktodak_api/common/security/component/oauth2/TodakOauth2UserDetailsService.java
@@ -1,4 +1,4 @@
-package com.heartsave.todaktodak_api.common.security.component;
+package com.heartsave.todaktodak_api.common.security.component.oauth2;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
